+++ packages/simulator/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @mossxyz/simulator
+# @themoss/simulator
 
 ## 0.1.0
 
@@ -29,4 +29,4 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @mossxyz/core@0.1.0
+  - @themoss/core@0.1.0